refactor: simplify notifications per PR feedback
- Rename feature folder: system-notifications -> notifications
- Rename component: SystemNotificationBanner -> Notifications
- Remove NotificationSettingsResponse DTO (use anonymous object)
- Remove force-refresh-coordinator (consumeSelfInitiatedFlag hack)
- Remove resolveDisplayMessage (over-engineered resolution)
- Remove MaxNotificationMessageLength validation (unwanted)
- Use useEventListener from runed (project pattern)
- Simplify banner: just show message or fallback like legacy Angular
- Add newline before return after control statements
- Revert .gitignore gstack addition
- Remove maxlength from textareas
