+++ .gitignore
@@ -84,4 +84,3 @@ debug-storybook.log
 .devcontainer/devcontainer-lock.json
 
 *.lscache
-.gstack/
