--- src/Exceptionless.Core/Models/Messaging/NotificationSettingsResponse.cs
@@ -1,7 +1,0 @@
-namespace Exceptionless.Core.Messaging.Models;
-
-public record NotificationSettingsResponse
-{
-    public string? ConfiguredSystemNotificationMessage { get; init; }
-    public SystemNotification? SystemNotification { get; init; }
-}
