+++ src/Exceptionless.Web/ClientApp/src/lib/features/notifications/api.svelte.ts
